Use time as a 64 bit integer for use in time.Unix()

+++ api/structsItem.go
@@ -12,7 +12,7 @@ type HNItem struct {
 	Parts       []int  `json:"parts"`
 	Score       int    `json:"score"`
 	Text        string `json:"text"`
-	Time        int    `json:"time"`
+	Time        int64  `json:"time"`
 	Title       string `json:"title"`
 	Type        string `json:"type"`
 	URL         string `json:"url"`
